Move GUI helper.
