+++ src/test/java/tempest/ui/pages/ManageModulesPageTest.java
@@ -6,6 +6,7 @@ import org.junit.Test;
 
 import tempest.State;
 import tempest.Supervisor;
+import tempest.helpers.GUIHelper;
 import tempest.ui.GUIManager;
 
 public class ManageModulesPageTest {
